moving elements to the helperFunctions js file to make them accessable all over
